requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.1
pylibdmtx==0.1.10
Werkzeug==3.1.6
gunicorn==21.2.0